move cli module to main
