+++ src/main.rs
@@ -1,4 +1,3 @@
-mod cli;
 mod config;
 mod error;
 mod pokemon;
@@ -7,8 +6,10 @@ use config::Config;
 use error::Error;
 use pokemon::*;
 
-use clap::{Args, Parser, Subcommand};
-use clap_complete::Shell;
+use clap::{arg, Args, Command, Parser, Subcommand};
+use clap_complete::{generate, Generator, Shell};
+use std::io;
+
 use rand::seq::SliceRandom;
 use rand::Rng;
 use rust_embed::RustEmbed;
@@ -227,13 +228,55 @@ fn show_random_pokemon(
     )
 }
 
+fn build_cli() -> Command {
+    let common_args = [
+        arg!(-i --info "Print pokedex entry (if it exists)"),
+        arg!(-s --shiny "Show the shiny pokemon version instead"),
+        arg!(--"no-title" "Do not display pokemon name"),
+        arg!(--"padding-left" "Set amount of padding to the left [default: 0]"),
+    ];
+    let init = Command::new("init")
+        .about("Generate shell completions")
+        .args([
+            arg!(["bash"]),
+            arg!(["zsh"]),
+            arg!(["fish"]),
+            arg!(["powershell"]),
+            arg!(["elvish"]),
+        ]);
+
+    let list = Command::new("list").about("List all names of pokemons");
+    let name = Command::new("name")
+        .about("Select pokemon by name: eg. 'pikachu'")
+        .arg(arg!([name] "Who's that pokemon!?"))
+        .args(&common_args);
+
+    let random = Command::new("random")
+        .about("Show random pokemon")
+        .arg(
+            arg!([GENERATIONS] "Generation number, range (1-9), or list of generations (1,3,6) [default: 1-9]"),
+        )
+        .args(common_args)
+        .args([
+            arg!(--"no-mega" "Do not show mega pokemon"),
+            arg!(--"no-gmax" "Do not show gigantamax pokemon"),
+            arg!(--"no-regional" "Do not show regional pokemon"),
+        ]);
+
+    Command::new("krabby").subcommands([init, list, name, random])
+}
+
+fn print_completions<G: Generator>(gen: G, cmd: &mut Command) {
+    generate(gen, cmd, cmd.get_name().to_string(), &mut io::stdout());
+}
+
 fn main() -> Result<(), Error> {
     let config = Config::load()?;
     let pokemon_db = Asset::get("pokemon.json").expect("Could not read pokemond db file");
     let pokemon = load_pokemon(&pokemon_db)?;
     let args = Cli::parse();
     match args.command {
-        Commands::Init(shell) => cli::print_completions(shell.shell, &mut cli::build()),
+        Commands::Init(shell) => print_completions(shell.shell, &mut build_cli()),
         Commands::List => list_pokemon_names(pokemon),
         Commands::Name(name) => show_pokemon_by_name(&name, pokemon, &config)?,
         Commands::Random(random) => show_random_pokemon(&random, pokemon, &config)?,
